enrichment-api/requirements.txt
uvicorn==0.18.2
watchfiles==0.15.0
websockets==10.3
requests==2.28.0
python-multipart==0.0.5
python-dotenv==0.20.0
python-jose==3.3.0
passlib==1.7.4
jsonschema==4.6.0
fastapi==0.79.0
colorlog==4.8.0
bcrypt==3.2.2